hexagon: merge htp-utils in htp-drv and remove unused code
