+++ ggml/src/ggml-hexagon/htp-drv.cpp
@@ -1,5 +1,9 @@
 // sample drv interface
 
+#pragma clang diagnostic ignored "-Wgnu-anonymous-struct"
+#pragma clang diagnostic ignored "-Wmissing-prototypes"
+#pragma clang diagnostic ignored "-Wsign-compare"
+
 #include <filesystem>
 #include <set>
 #include <sstream>
@@ -18,6 +22,8 @@
 #include "ggml-backend-dl.h"
 #include "ggml-impl.h"
 #include "htp-drv.h"
+
+#include <domain.h>
 
 //
 // Driver API types
@@ -95,10 +101,10 @@ void * rpcmem_alloc(int heapid, uint32_t flags, int size) {
 }
 
 void * rpcmem_alloc2(int heapid, uint32_t flags, size_t size) {
-    // fall back to rpcmem_alloc if v2 not supported
-    if (rpcmem_alloc2) {
+    if (rpcmem_alloc2_pfn) {
         return rpcmem_alloc2_pfn(heapid, flags, size);
     } else {
+        GGML_LOG_INFO("ggml-hex: rpcmem_alloc2 not found, falling back to rpcmem_alloc\n");
         return rpcmem_alloc_pfn(heapid, flags, size);
     }
 }
@@ -197,7 +203,7 @@ static std::string wstr_to_str(std::wstring_view wstr) {
                                             wstr.data(), (int) wstr.size(),
                                             nullptr, 0, nullptr, nullptr);
     if (bytes_needed == 0) {
-        GGML_LOG_ERROR("WideCharToMultiByte failed. Error %lu\n", GetLastError());
+        GGML_LOG_ERROR("ggml-hex: WideCharToMultiByte failed. Error %lu\n", GetLastError());
         throw std::runtime_error("Invalid wstring input");
     }
 
@@ -207,7 +213,7 @@ static std::string wstr_to_str(std::wstring_view wstr) {
                                             result.data(), bytes_needed,
                                             nullptr, nullptr);
     if (bytes_written == 0) {
-        GGML_LOG_ERROR("WideCharToMultiByte failed. Error %lu\n", GetLastError());
+        GGML_LOG_ERROR("ggml-hex: WideCharToMultiByte failed. Error %lu\n", GetLastError());
         throw std::runtime_error("Wstring conversion failed");
     }
     return result;
@@ -220,7 +226,7 @@ static std::string get_driver_path() {
     // Get a handle to the SCM database.
     SC_HANDLE schSCManager = OpenSCManagerW(NULL, NULL, STANDARD_RIGHTS_READ);
     if (nullptr == schSCManager) {
-        GGML_LOG_ERROR("Failed to open SCManager. Error: %lu\n", GetLastError());
+        GGML_LOG_ERROR("ggml-hex: Failed to open SCManager. Error: %lu\n", GetLastError());
         return result;
     }
 
@@ -230,7 +236,7 @@ static std::string get_driver_path() {
                                         SERVICE_QUERY_CONFIG);  // need query config access
 
     if (nullptr == schService) {
-        GGML_LOG_ERROR("Failed to open qcnspmcdm service. Error: %lu", GetLastError());
+        GGML_LOG_ERROR("ggml-hex: Failed to open qcnspmcdm service. Error: %lu", GetLastError());
         CloseServiceHandle(schSCManager);
         return result;
     }
@@ -239,7 +245,7 @@ static std::string get_driver_path() {
     DWORD bufferSize; 
     if (!QueryServiceConfigW(schService, NULL, 0, &bufferSize) &&
         (GetLastError() != ERROR_INSUFFICIENT_BUFFER)) {
-        GGML_LOG_ERROR("Failed to query service config. Error: %lu\n", GetLastError());
+        GGML_LOG_ERROR("ggml-hex: Failed to query service config. Error: %lu\n", GetLastError());
         CloseServiceHandle(schService);
         CloseServiceHandle(schSCManager);
         return result;
@@ -248,7 +254,7 @@ static std::string get_driver_path() {
     LPQUERY_SERVICE_CONFIGW serviceConfig =
         static_cast<LPQUERY_SERVICE_CONFIGW>(LocalAlloc(LMEM_FIXED, bufferSize));
     if (!QueryServiceConfigW(schService, serviceConfig, bufferSize, &bufferSize)) {
-        fprintf(stderr, "Failed to query service config. Error: %lu\n", GetLastError());
+        fprintf(stderr, "ggml-hex: Failed to query service config. Error: %lu\n", GetLastError());
         LocalFree(serviceConfig);
         CloseServiceHandle(schService);
         CloseServiceHandle(schSCManager);
@@ -269,7 +275,7 @@ static std::string get_driver_path() {
     // "\SystemRoot" should be replace with a correct one (e.g. C:\Windows)
     const std::wstring systemRootPlaceholder = L"\\SystemRoot";
     if (0 != driverPath.compare(0, systemRootPlaceholder.length(), systemRootPlaceholder)) {
-        GGML_LOG_ERROR("String pattern not found in driver path.\n");
+        GGML_LOG_ERROR("ggml-hex: String pattern not found in driver path.\n");
         return result;
     }
 
@@ -279,7 +285,7 @@ static std::string get_driver_path() {
     // Query the number of wide charactors this variable requires
     DWORD numWords = GetEnvironmentVariableW(systemRootEnv.c_str(), NULL, 0);
     if (numWords == 0) {
-        GGML_LOG_ERROR("Failed get systemRoot environment variable\n");
+        GGML_LOG_ERROR("ggml-hex: Failed get systemRoot environment variable\n");
         return result;
     }
 
@@ -287,7 +293,7 @@ static std::string get_driver_path() {
     std::vector<wchar_t> systemRoot(numWords + 1);
     numWords = GetEnvironmentVariableW(systemRootEnv.c_str(), systemRoot.data(), numWords + 1);
     if (numWords == 0) {
-        GGML_LOG_ERROR("Failed to read windir environment variable\n");
+        GGML_LOG_ERROR("ggml-hex: Failed to read windir environment variable\n");
         return result;
     }
     driverPath.replace(0, systemRootPlaceholder.length(), std::wstring(systemRoot.data()));
@@ -308,24 +314,23 @@ int htpdrv_init() {
     std::string drv_path = "libcdsprpc.so";
 #endif
     if (initialized) {
-        GGML_LOG_INFO("%s: Driver already loaded\n", __func__);
+        GGML_LOG_INFO("ggml-hex: Driver already loaded\n");
         return AEE_SUCCESS;
     }
-    GGML_LOG_INFO("%s: Loading driver %s\n", __func__, drv_path.c_str());
+    GGML_LOG_INFO("ggml-hex: Loading driver %s\n", drv_path.c_str());
 
     fs::path path{ drv_path.c_str() };
     dl_handle_ptr handle { dl_load_library(path) };
     if (!handle) {
-        GGML_LOG_ERROR("%s: failed to load %s: %s\n",
-                       __func__, path.u8string().c_str(), dl_error());
+        GGML_LOG_ERROR("ggml-hex: failed to load %s: %s\n", path.u8string().c_str(), dl_error());
         return AEE_EUNABLETOLOAD;
     }
 
 #define dlsym(drv, type, pfn, symbol, ignore)                               \
     do {                                                                    \
         pfn = (type) dl_get_sym(drv, #symbol);                              \
-        if (!ignore && nullptr == pfn) {                                               \
-            GGML_LOG_ERROR("%s: failed to dlsym %s\n", __func__, #symbol);  \
+        if (!ignore && nullptr == pfn) {                                    \
+            GGML_LOG_ERROR("ggml-hex: failed to dlsym %s\n", #symbol);      \
             return AEE_EUNABLETOLOAD;                                       \
         }                                                                   \
     } while (0)
@@ -352,4 +357,62 @@ int htpdrv_init() {
     initialized         = true;
 
     return AEE_SUCCESS;
+}
+
+domain * get_domain(int domain_id) {
+    int i    = 0;
+    int size = sizeof(supported_domains) / sizeof(domain);
+
+    for (i = 0; i < size; i++) {
+        if (supported_domains[i].id == domain_id) {
+            return &supported_domains[i];
+        }
+    }
+
+    return NULL;
+}
+
+int get_hex_arch_ver(int domain, int * arch) {
+    if (!remote_handle_control_pfn) {
+        GGML_LOG_ERROR("ggml-hex: remote_handle_control is not supported on this device\n");
+        return AEE_EUNSUPPORTEDAPI;
+    }
+
+    struct remote_dsp_capability arch_ver;
+    arch_ver.domain       = (uint32_t) domain;
+    arch_ver.attribute_ID = ARCH_VER;
+    arch_ver.capability   = (uint32_t) 0;
+
+    int err = remote_handle_control(DSPRPC_GET_DSP_INFO, &arch_ver, sizeof(arch_ver));
+    if ((err & 0xff) == (AEE_EUNSUPPORTEDAPI & 0xff)) {
+        GGML_LOG_ERROR("ggml-hex: FastRPC capability API is not supported on this device\n");
+        return AEE_EUNSUPPORTEDAPI;
+    }
+
+    if (err != AEE_SUCCESS) {
+        GGML_LOG_ERROR("ggml-hex: FastRPC capability query failed (err %d)\n", err);
+        return err;
+    }
+
+    switch (arch_ver.capability & 0xff) {
+        case 0x68:
+            *arch = 68;
+            return 0;
+        case 0x69:
+            *arch = 69;
+            return 0;
+        case 0x73:
+            *arch = 73;
+            return 0;
+        case 0x75:
+            *arch = 75;
+            return 0;
+        case 0x79:
+            *arch = 79;
+            return 0;
+        case 0x81:
+            *arch = 81;
+            return 0;
+    }
+    return -1;
 }
